AndroidX support + setServerUrl addition

+++ ios/ANEAmplitude/ANEAmplitude/FRETypeUtils.m
@@ -18,7 +18,7 @@ FREResult FREGetObjectAsString( FREObject object, NSString** value )
 FREResult FRENewObjectFromString( NSString* string, FREObject* asString )
 {
     const char* utf8String = string.UTF8String;
-    unsigned long length = strlen( utf8String );
+    uint32_t length = (uint32_t)strlen( utf8String );
     return FRENewObjectFromUTF8( length + 1, (uint8_t*) utf8String, asString );
 }
